requirements.txt
langchain
langchain-core
langchain-openai
langchain-groq
langchain_huggingface
langchain-community
langgraph
groq
openai
python-dotenv
faiss-cpu
tiktoken
fastapi
uvicorn
requests